add backup check

+++ restore.sh
@@ -5,19 +5,30 @@
 # =======================================================================================================================
 
 NEXTCLOUD_DIR="/var/www/html/nextcloud"
+BCK_HOST="backup"
 
+if ssh backup '[ ! -d /root/nextcloud_* ]'
+then
+    echo "No backup present on the server, please make a backup before using this script" >&2
+    exit 1
+fi
+
+# I also like to live dangerously
 rm -r $NEXTCLOUD_DIR
 
 # Drop previous database
 mysql -u root -proot -e "DROP DATABASE nextcloud"
 
 # get last backup
-LAST_DIR=$(ssh backup "ls -t /root/backups/nextcloud/ | head -n 1")
+LAST_DIR=$(ssh backup "ls -td */ /root/ | head -n 1")
 
-rsync -Aavxziptgo backup:/root/backups/nextcloud/$LAST_DIR /var/www/html/
+# sync
+rsync -Aavxziptgo $BCK_HOST:/root/$LAST_DIR /var/www/html/
 
+# rename into "nextcloud"
 mv /var/www/html/$LAST_DIR $NEXTCLOUD_DIR
 
+# Turn maintenance mode ON
 sudo -u www-data php $NEXTCLOUD_DIR/occ maintenance:mode --on
 
 # Create new database
